Возможность менять позицию досок и колонок

+++ src/main/java/kg/core/task/dtos/UpdatePosition.java
@@ -4,6 +4,6 @@ import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(description = "DTO для изменения позиции задачи")
 public record UpdatePosition(
-        @Schema(description = "Позиция задачи") Long position
+        @Schema(description = "Позиция задачи") Integer position
 ) {
 }
